Split types.go into several files
No other changes happened, except for adding imports.
