+++ internal/campaigns/changeset_spec_description.go
@@ -1,0 +1,110 @@
+package campaigns
+
+import (
+	"github.com/graph-gophers/graphql-go"
+	"github.com/pkg/errors"
+)
+
+type ChangesetSpecDescription struct {
+	BaseRepository graphql.ID `json:"baseRepository,omitempty"`
+
+	// If this is not empty, the description is a reference to an existing
+	// changeset and the rest of these fields are empty.
+	ExternalID string `json:"externalID,omitempty"`
+
+	BaseRev string `json:"baseRev,omitempty"`
+	BaseRef string `json:"baseRef,omitempty"`
+
+	HeadRepository graphql.ID `json:"headRepository,omitempty"`
+	HeadRef        string     `json:"headRef,omitempty"`
+
+	Title string `json:"title,omitempty"`
+	Body  string `json:"body,omitempty"`
+
+	Commits []GitCommitDescription `json:"commits,omitempty"`
+
+	Published bool `json:"published,omitempty"`
+}
+
+// Type returns the ChangesetSpecDescriptionType of the ChangesetSpecDescription.
+func (d *ChangesetSpecDescription) Type() ChangesetSpecDescriptionType {
+	if d.ExternalID != "" {
+		return ChangesetSpecDescriptionTypeExisting
+	}
+	return ChangesetSpecDescriptionTypeBranch
+}
+
+// IsExisting returns whether the description is of type
+// ChangesetSpecDescriptionTypeExisting.
+func (d *ChangesetSpecDescription) IsImportingExisting() bool {
+	return d.Type() == ChangesetSpecDescriptionTypeExisting
+}
+
+// IsBranch returns whether the description is of type
+// ChangesetSpecDescriptionTypeBranch.
+func (d *ChangesetSpecDescription) IsBranch() bool {
+	return d.Type() == ChangesetSpecDescriptionTypeBranch
+}
+
+// ChangesetSpecDescriptionType tells the consumer what the type of a
+// ChangesetSpecDescription is without having to look into the description.
+// Useful in the GraphQL when a HiddenChangesetSpec is returned.
+type ChangesetSpecDescriptionType string
+
+// Valid ChangesetSpecDescriptionTypes kinds
+const (
+	ChangesetSpecDescriptionTypeExisting ChangesetSpecDescriptionType = "EXISTING"
+	ChangesetSpecDescriptionTypeBranch   ChangesetSpecDescriptionType = "BRANCH"
+)
+
+// ErrNoCommits is returned by (*ChangesetSpecDescription).Diff if the
+// description doesn't have any commits descriptions.
+var ErrNoCommits = errors.New("changeset description doesn't contain commit descriptions")
+
+// Diff returns the Diff of the first GitCommitDescription in Commits. If the
+// ChangesetSpecDescription doesn't have Commits it returns ErrNoCommits.
+//
+// We currently only support a single commit in Commits. Once we support more,
+// this method will need to be revisited.
+func (d *ChangesetSpecDescription) Diff() (string, error) {
+	if len(d.Commits) == 0 {
+		return "", ErrNoCommits
+	}
+	return d.Commits[0].Diff, nil
+}
+
+// CommitMessage returns the Message of the first GitCommitDescription in Commits. If the
+// ChangesetSpecDescription doesn't have Commits it returns ErrNoCommits.
+//
+// We currently only support a single commit in Commits. Once we support more,
+// this method will need to be revisited.
+func (d *ChangesetSpecDescription) CommitMessage() (string, error) {
+	if len(d.Commits) == 0 {
+		return "", ErrNoCommits
+	}
+	return d.Commits[0].Message, nil
+}
+
+// AuthorName returns the author name of the first GitCommitDescription in Commits. If the
+// ChangesetSpecDescription doesn't have Commits it returns ErrNoCommits.
+//
+// We currently only support a single commit in Commits. Once we support more,
+// this method will need to be revisited.
+func (d *ChangesetSpecDescription) AuthorName() (string, error) {
+	if len(d.Commits) == 0 {
+		return "", ErrNoCommits
+	}
+	return d.Commits[0].AuthorName, nil
+}
+
+// AuthorEmail returns the author email of the first GitCommitDescription in Commits. If the
+// ChangesetSpecDescription doesn't have Commits it returns ErrNoCommits.
+//
+// We currently only support a single commit in Commits. Once we support more,
+// this method will need to be revisited.
+func (d *ChangesetSpecDescription) AuthorEmail() (string, error) {
+	if len(d.Commits) == 0 {
+		return "", ErrNoCommits
+	}
+	return d.Commits[0].AuthorEmail, nil
+}
